+++ dotfiles/.config/ml4w/themes/modern/theme.sh
@@ -1,0 +1,28 @@
+#!/usr/bin/env bash
+# ML4W Theme Modern
+
+# Set waybar
+echo "/ml4w-modern;/ml4w-modern/default" > $HOME/.config/ml4w/settings/waybar-theme.sh
+$HOME/.config/waybar/launch.sh &
+
+# Set nwg-dock-hyprland
+echo "modern" > $HOME/.config/ml4w/settings/dock-theme
+$HOME/.config/nwg-dock-hyprland/launch.sh &
+
+# Set swaync
+echo '@import "themes/modern/style.css";' > $HOME/.config/swaync/style.css
+swaync-client -rs
+
+# Set wlogout
+echo '@import "themes/modern/style.css";' > $HOME/.config/wlogout/style.css
+
+# Set launcher
+echo 'walker' > $HOME/.config/ml4w/settings/launcher
+
+# Set Window Border
+echo 'source = ~/.config/hypr/conf/windows/default.conf' > $HOME/.config/hypr/conf/window.conf
+
+# Set rofi
+echo '* { border-width: 2px; }' > $HOME/.config/ml4w/settings/rofi-border.rasi
+
+echo ":: Theme set to Modern"